+++ frontend/containers/open-call-page/open-call-chart/index.ts
@@ -1,0 +1,3 @@
+export { default } from './component';
+
+export type { OpenCallChartProps } from './types';
