+++ frontend/containers/open-call-page/open-call-chart/types.ts
@@ -1,0 +1,7 @@
+export type OpenCallChartProps = {
+  openCallRange: {
+    consumed: number;
+    remaining: number;
+    deadline: string;
+  };
+};
