+++ frontend/containers/open-call-page/open-call-header/index.ts
@@ -1,0 +1,2 @@
+export { default } from './component';
+export type { OpenCallHeaderProps } from './types';
