+++ frontend/containers/open-call-page/open-call-header/types.ts
@@ -1,0 +1,7 @@
+import { Enum } from 'types/enums';
+import { OpenCall } from 'types/open-calls';
+
+export type OpenCallHeaderProps = {
+  openCall: OpenCall;
+  instrumentTypeEnums: Enum[];
+};
